relocate function; fix issue of updating rider
