+++ config/rbac/role.yaml
@@ -72,3 +72,7 @@ rules:
       - 'get'
       - 'list'
       - 'watch'
+  - apiGroups: ["autoscaling.k8s.io"]
+    resources:
+      - verticalpodautoscalers
+    verbs: ["*"]
