+++ internal/controller/common/rollout_controller.go
@@ -2,13 +2,11 @@ package common
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"math"
 	"strconv"
 	"strings"
 
-	fasttemplate "github.com/valyala/fasttemplate"
 	k8stypes "k8s.io/apimachinery/pkg/types"
 
 	"github.com/numaproj/numaplane/internal/common"
@@ -262,30 +260,4 @@ func GetChildName(ctx context.Context, rolloutObject RolloutObject, controller R
 	} else {
 		return existingChild.GetName(), nil
 	}
-}
-
-// resolves templated definitions of a resource with any arguments
-func ResolveTemplateSpec(data any, args map[string]interface{}) (map[string]interface{}, error) {
-
-	// marshal data to cast as a string
-	dataBytes, err := json.Marshal(data)
-	if err != nil {
-		return nil, err
-	}
-
-	// create and execute template with supplied arguments
-	tmpl, err := fasttemplate.NewTemplate(string(dataBytes), "{{", "}}")
-	if err != nil {
-		return nil, err
-	}
-	templatedSpec := tmpl.ExecuteString(args)
-
-	// unmarshal into map to be returned and used for resource spec
-	var resolvedTmpl map[string]interface{}
-	err = json.Unmarshal([]byte(templatedSpec), &resolvedTmpl)
-	if err != nil {
-		return nil, err
-	}
-
-	return resolvedTmpl, nil
 }
