+++ internal/util/util_test.go
@@ -316,3 +316,100 @@ func Test_ToInt64(t *testing.T) {
 		})
 	}
 }
+
+// This tests the ResolveTemplateSpec function
+// This function resolves any templated definition with the provided arguments
+func TestResolveTemplateSpec(t *testing.T) {
+
+	tests := []struct {
+		name           string
+		data           any
+		args           map[string]interface{}
+		expectedOutput map[string]interface{}
+	}{
+		{
+			name: "standard test",
+			data: map[string]interface{}{
+				"spec": map[string]interface{}{
+					"configMap": map[string]interface{}{
+						"name":      "test-volume-{{.monovertex-name}}",
+						"namespace": "{{.monovertex-namespace}}",
+					},
+				},
+			},
+			args: map[string]interface{}{
+				".monovertex-name":      "my-monovertex-0",
+				".monovertex-namespace": "test-namespace",
+			},
+			expectedOutput: map[string]interface{}{
+				"spec": map[string]interface{}{
+					"configMap": map[string]interface{}{
+						"name":      "test-volume-my-monovertex-0",
+						"namespace": "test-namespace",
+					},
+				},
+			},
+		},
+		{
+			name: "argument is used multiple times",
+			data: map[string]interface{}{
+				"annotations": map[string]interface{}{
+					"link.argocd.argoproj.io/external-link": "https://splunk.intuit.com/en-US/app/search/search?q=search%20(index%3Daccounting-ledger%20source%3Diks2%2Fsbg-qbo-ppd-usw2-k8s%2F*%2F{{.monovertex-namespace}}%2F*)%20OR%20(index%3Diks%20source%3Diks2%2Fsbg-qbo-ppd-usw2-k8s%2F*%2F{{.monovertex-namespace}}%2F*)%20kubernetes_pod%3D{{.monovertex-name}}*&display.page.search.mode=fast&dispatch.sample_ratio=1&earliest=-5m&latest=now",
+				},
+				"spec": map[string]interface{}{
+					"configMap": map[string]interface{}{
+						"name":      "test-volume-{{.monovertex-name}}",
+						"namespace": "{{.monovertex-namespace}}",
+					},
+				},
+			},
+			args: map[string]interface{}{
+				".monovertex-name":      "my-monovertex-0",
+				".monovertex-namespace": "test-namespace",
+			},
+			expectedOutput: map[string]interface{}{
+				"annotations": map[string]interface{}{
+					"link.argocd.argoproj.io/external-link": "https://splunk.intuit.com/en-US/app/search/search?q=search%20(index%3Daccounting-ledger%20source%3Diks2%2Fsbg-qbo-ppd-usw2-k8s%2F*%2Ftest-namespace%2F*)%20OR%20(index%3Diks%20source%3Diks2%2Fsbg-qbo-ppd-usw2-k8s%2F*%2Ftest-namespace%2F*)%20kubernetes_pod%3Dmy-monovertex-0*&display.page.search.mode=fast&dispatch.sample_ratio=1&earliest=-5m&latest=now",
+				},
+				"spec": map[string]interface{}{
+					"configMap": map[string]interface{}{
+						"name":      "test-volume-my-monovertex-0",
+						"namespace": "test-namespace",
+					},
+				},
+			},
+		},
+		{
+			name: "input is not templated",
+			data: map[string]interface{}{
+				"spec": map[string]interface{}{
+					"configMap": map[string]interface{}{
+						"name":      "test-volume-my-monovertex-0",
+						"namespace": "test-namespace",
+					},
+				},
+			},
+			args: map[string]interface{}{
+				".monovertex-name":      "my-monovertex-0",
+				".monovertex-namespace": "test-namespace",
+			},
+			expectedOutput: map[string]interface{}{
+				"spec": map[string]interface{}{
+					"configMap": map[string]interface{}{
+						"name":      "test-volume-my-monovertex-0",
+						"namespace": "test-namespace",
+					},
+				},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result, err := ResolveTemplateSpec(tt.data, tt.args)
+			assert.NoError(t, err)
+			assert.Equal(t, tt.expectedOutput, result)
+		})
+	}
+
+}
